Separate perm backend and manager tests
